Project 2: Magic Square
A magic square is a grid of unique numbers (i.e. no number can appear
more than once) in which the sum of all the rows, columns, and diagonals
are the same. Here is an example in which all of the rows, columns and
diagonals (upper left corner, center, lower right corner and upper right
corner, center, and lower left corner) sum to 15. Generally, the numbers
within a magic square range between 1 and n^2, where n is the size of the
side of the square.